add type files
