+++ packages/core/src/core/ChatProcessor.ts
@@ -1,6 +1,6 @@
 import { ChatService } from '../services/chat/ChatService';
 import { MemoryManager } from './MemoryManager';
-import { Message, MessageWithVision, ChatType, Screenplay } from '../types';
+import { Message, MessageWithVision, ChatType } from '../types';
 import { EventEmitter } from './EventEmitter';
 import { textsToScreenplay } from '../utils/screenplay';
 import { DEFAULT_VISION_PROMPT } from '../constants/prompts';
