+++ packages/core/src/types/voiceEngine.ts
@@ -1,0 +1,1 @@
+export type voiceEngineType = 'voicevox' | 'voicepeak' | 'openai' | 'nijivoice' | 'aivisSpeech';
